Cleanup old compass files, add test command
