+++ src/main/java/appeng/core/localization/PlayerMessages.java
@@ -45,6 +45,8 @@ public enum PlayerMessages {
     InterfaceInOtherDimParam,
     InterfaceInOtherDim,
     InterfaceHighlighted,
+
+    CompassTestSection
     ;
 
     public ITextComponent get(Object... params) {
