--- src/main/java/appeng/services/compass/ICompassCallback.java
@@ -1,34 +1,0 @@
-/*
- * This file is part of Applied Energistics 2.
- * Copyright (c) 2013 - 2014, AlgorithmX2, All rights reserved.
- *
- * Applied Energistics 2 is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Applied Energistics 2 is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with Applied Energistics 2.  If not, see <http://www.gnu.org/licenses/lgpl>.
- */
-
-package appeng.services.compass;
-
-
-@Deprecated
-public interface ICompassCallback {
-
-    /**
-     * Called from another thread.
-     *
-     * @param hasResult true if found a target
-     * @param spin      true if should spin
-     * @param radians   radians
-     * @param dist      distance
-     */
-    void calculatedDirection(boolean hasResult, boolean spin, double radians, double dist);
-}
