+++ src/main/resources/assets/appliedenergistics2/lang/en_us.lang
@@ -93,6 +93,7 @@ chat.appliedenergistics2.isNowUnlocked=Monitor is now Unlocked.
 chat.appliedenergistics2.InterfaceInOtherDimParam=Interface located at dimension: %d [ %s ] and cannot be highlighted
 chat.appliedenergistics2.InterfaceInOtherDim=Interface is located in another dimension and cannot be highlighted
 chat.appliedenergistics2.InterfaceHighlighted=The interface is now highlighted at X: %s Y: %s Z: %s
+chat.appliedenergistics2.CompassTestSection=Section [y=%d-%d] %d: %s
 
 // Creative Tabs
 itemGroup.appliedenergistics2=Applied Energistics 2
@@ -642,6 +643,7 @@ commands.ae2.ChunkLogger=Toggles Chunk Loading and unloading to the server log. 
 commands.ae2.ChunkLoggerOn=Chunk Logging is now on
 commands.ae2.ChunkLoggerOff=Chunk Logging is now off
 commands.ae2.Supporters=Displays a list of AE2 Supporters
+commands.ae2.Compass=Tests each Section in the current chunk for compass target. ( OP )
 
 // Achievements
 achievement.ae2.Root=Applied Energistics
